sekura/requirements.txt
django >= 4.0
django-rest-framework
djangorestframework-simplejwt
psycopg2
pytz
uvicorn
gunicorn
watchgod
inotify
pywatchman
redis
hiredis
celery
django-celery-results
django-celery-beat
flower
